refactored testing
